requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
shap
joblib
fpdf2
anthropic
plotly
